+++ src/components/consumerRegistry/DeleteConsumer.js
@@ -1,14 +1,13 @@
 import { useState } from "react";
 import { ethers } from "ethers";
-import "../css/bootstrap.css";
-import ConsumerRegistry from "../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
+import "../../css/bootstrap.css";
+import ConsumerRegistry from "../../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
 
 // May need to pdate on deployment. This is the address the contract is deployed to.\
 const consumerRegistryAddress = process.env.REACT_APP_CONSUMER_ADDRESS;
 
-const VerifyAttestationWithSig = () => {
+const DeleteConsumer = () => {
   const [did, setDid] = useState("");
-  const [attestationSig, setAttestationSig] = useState("");
   const [result, setResult] = useState("");
 
   // uses metamask injected browser window to make sure consumer has a connected account
@@ -17,34 +16,34 @@ const VerifyAttestationWithSig = () => {
   }
 
   // call to the insert method of the smart contract
-  async function verify() {
+  async function deleteConsumer() {
     // making sure input is not empty
-    if (!did || !attestationSig) {
+    if (!did) {
       console.log("Insert values are empty");
       setResult(`Insert Values are empty`);
       return;
     }
-
     if (typeof window.ethereum !== "undefined") {
       await requestAccount();
       const provider = new ethers.providers.Web3Provider(window.ethereum);
-
+      // we need a signer because insert requires a transaction
+      const signer = provider.getSigner();
       const contract = new ethers.Contract(
         consumerRegistryAddress,
         ConsumerRegistry.abi,
-        provider
+        signer
       );
 
+      // Listening for the emmitted event
+      contract.on("EntryDeleted", (did) => {
+        setResult(`Event caught. Consumer deleted with did: ${did}`);
+      });
+
       try {
-        const data = await contract.verifyAttestationWithSig(
-          did,
-          attestationSig
-        );
-        if (data) {
-          setResult(`Function returned true`);
-        } else {
-          setResult(`Function returned false`);
-        }
+        const transaction = await contract.deleteConsumer(did); //is there a way to get return value
+        // of non view function?
+        await transaction.wait();
+        // console.log({ transaction });
       } catch (err) {
         console.log("Error: ", err);
         setResult("Error. Check console");
@@ -55,25 +54,20 @@ const VerifyAttestationWithSig = () => {
   return (
     <div className="App">
       <header className="App-header">
-        <h2>Verify consumer, need to fix</h2>
+        <h2>Delete Consumer Entry</h2>
         <input
           type="text"
           required
-          placeholder="DID"
+          placeholder="Set DID"
           onChange={(e) => setDid(e.target.value)}
         />
-        <input
-          type="text"
-          required
-          placeholder="Attestation Signature"
-          onChange={(e) => setAttestationSig(e.target.value)}
-        />
-        <button className="btn btn-outline-secondary" onClick={verify}>
-          Verify Consumer
+        <button className="btn btn-outline-secondary" onClick={deleteConsumer}>
+          Delete Entry
         </button>
+        {result}
       </header>
     </div>
   );
 };
 
-export default VerifyAttestationWithSig;
+export default DeleteConsumer;
