+++ src/components/consumerRegistry/GetConsumer.js
@@ -1,20 +1,28 @@
 import { useState } from "react";
 import { ethers } from "ethers";
-import "../css/bootstrap.css";
+import "../../css/bootstrap.css";
 
-import ConsumerRegistry from "../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
+import ConsumerRegistry from "../../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
 
 const consumerRegistryAddress = process.env.REACT_APP_CONSUMER_ADDRESS;
 
-const TotalConsumers = () => {
+const GetConsumer = () => {
+  const [did, setDid] = useState("");
   const [result, setResult] = useState("");
+
   // uses metamask injected browser window to make sure consumer has a connected account
   async function requestAccount() {
     await window.ethereum.request({ method: "eth_requestAccounts" });
   }
 
   // call to the insert method of the smart contract
-  async function getTotalEntries() {
+  async function getConsumer() {
+    // making sure input is not empty
+    if (!did) {
+      console.log("Insert values are empty");
+      setResult(`Insert Values are empty`);
+      return;
+    }
     if (typeof window.ethereum !== "undefined") {
       await requestAccount();
       const provider = new ethers.providers.Web3Provider(window.ethereum);
@@ -25,9 +33,9 @@ const TotalConsumers = () => {
         provider
       );
       try {
-        const data = await contract.getTotalEntries();
+        const data = await contract.getConsumer(did);
         console.log({ data });
-        setResult(`Total Entries: ${data}`);
+        setResult(`Retrieved Did: ${data[0]} with Key: ${data[1]}`);
       } catch (err) {
         console.log("Error: ", err);
         setResult("Error. Check console");
@@ -38,9 +46,15 @@ const TotalConsumers = () => {
   return (
     <div className="App">
       <header className="App-header">
-        <h2>Get Total Entries</h2>
-        <button className="btn btn-outline-secondary" onClick={getTotalEntries}>
-          Entry Count
+        <h2>Get Consumer Entry With Consumer DID</h2>
+        <input
+          type="text"
+          required
+          placeholder="DID"
+          onChange={(e) => setDid(e.target.value)}
+        />
+        <button className="btn btn-outline-secondary" onClick={getConsumer}>
+          Get Consumer
         </button>
         {result}
       </header>
@@ -48,4 +62,4 @@ const TotalConsumers = () => {
   );
 };
 
-export default TotalConsumers;
+export default GetConsumer;
