+++ src/components/consumerRegistry/VerifyAttestationWithSig.js
@@ -1,13 +1,14 @@
 import { useState } from "react";
 import { ethers } from "ethers";
-import "../css/bootstrap.css";
-import ConsumerRegistry from "../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
+import "../../css/bootstrap.css";
+import ConsumerRegistry from "../../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
 
 // May need to pdate on deployment. This is the address the contract is deployed to.\
 const consumerRegistryAddress = process.env.REACT_APP_CONSUMER_ADDRESS;
 
-const DeleteConsumer = () => {
+const VerifyAttestationWithSig = () => {
   const [did, setDid] = useState("");
+  const [attestationSig, setAttestationSig] = useState("");
   const [result, setResult] = useState("");
 
   // uses metamask injected browser window to make sure consumer has a connected account
@@ -16,34 +17,34 @@ const DeleteConsumer = () => {
   }
 
   // call to the insert method of the smart contract
-  async function deleteConsumer() {
+  async function verify() {
     // making sure input is not empty
-    if (!did) {
+    if (!did || !attestationSig) {
       console.log("Insert values are empty");
       setResult(`Insert Values are empty`);
       return;
     }
+
     if (typeof window.ethereum !== "undefined") {
       await requestAccount();
       const provider = new ethers.providers.Web3Provider(window.ethereum);
-      // we need a signer because insert requires a transaction
-      const signer = provider.getSigner();
+
       const contract = new ethers.Contract(
         consumerRegistryAddress,
         ConsumerRegistry.abi,
-        signer
+        provider
       );
 
-      // Listening for the emmitted event
-      contract.on("EntryDeleted", (did) => {
-        setResult(`Event caught. Consumer deleted with did: ${did}`);
-      });
-
       try {
-        const transaction = await contract.deleteConsumer(did); //is there a way to get return value
-        // of non view function?
-        await transaction.wait();
-        // console.log({ transaction });
+        const data = await contract.verifyAttestationWithSig(
+          did,
+          attestationSig
+        );
+        if (data) {
+          setResult(`Function returned true`);
+        } else {
+          setResult(`Function returned false`);
+        }
       } catch (err) {
         console.log("Error: ", err);
         setResult("Error. Check console");
@@ -54,20 +55,25 @@ const DeleteConsumer = () => {
   return (
     <div className="App">
       <header className="App-header">
-        <h2>Delete Consumer Entry</h2>
+        <h2>Verify consumer, need to fix</h2>
         <input
           type="text"
           required
-          placeholder="Set DID"
+          placeholder="DID"
           onChange={(e) => setDid(e.target.value)}
         />
-        <button className="btn btn-outline-secondary" onClick={deleteConsumer}>
-          Delete Entry
+        <input
+          type="text"
+          required
+          placeholder="Attestation Signature"
+          onChange={(e) => setAttestationSig(e.target.value)}
+        />
+        <button className="btn btn-outline-secondary" onClick={verify}>
+          Verify Consumer
         </button>
-        {result}
       </header>
     </div>
   );
 };
 
-export default DeleteConsumer;
+export default VerifyAttestationWithSig;
