+++ src/components/index.js
@@ -1,11 +1,13 @@
-export { default as Insert } from "./Insert";
-export { default as DeleteEntry } from "./DeleteEntry";
-export { default as GetEntry } from "./GetEntry";
-export { default as GetTotalEntries } from "./GetTotalEntries";
-export { default as UpdateEntry } from "./UpdateEntry";
-export { default as InsertConsumer } from "./InsertConsumer";
-export { default as GetConsumer } from "./GetConsumer";
-export { default as UpdateConsumer } from "./UpdateConsumer";
-export { default as DeleteConsumer } from "./DeleteConsumer";
-export { default as VerifyAttestationWithSig} from "./VerifyAttestationWithSig"
-export { default as TotalConsumers} from "./TotalConsumers"
+export { default as Insert } from "./userRegistry/Insert";
+export { default as DeleteEntry } from "./userRegistry/DeleteEntry";
+export { default as GetEntry } from "./userRegistry/GetEntry";
+export { default as GetTotalEntries } from "./userRegistry/GetTotalEntries";
+export { default as UpdateEntry } from "./userRegistry/UpdateEntry";
+export { default as InsertConsumer } from "./consumerRegistry/InsertConsumer";
+export { default as GetConsumer } from "./consumerRegistry/GetConsumer";
+export { default as UpdateConsumer } from "./consumerRegistry/UpdateConsumer";
+export { default as DeleteConsumer } from "./consumerRegistry/DeleteConsumer";
+export { default as VerifyAttestationWithSig } from "./consumerRegistry/VerifyAttestationWithSig";
+export { default as TotalConsumers } from "./consumerRegistry/TotalConsumers";
+
+export { default as QRCode } from "./userRegistry/qrcode";
