+++ src/components/userRegistry/GetTotalEntries.js
@@ -1,28 +1,20 @@
 import { useState } from "react";
 import { ethers } from "ethers";
-import "../css/bootstrap.css";
+import "../../css/bootstrap.css";
 
-import UserRegistry from "../artifacts/contracts/UserRegistry.sol/UserRegistry.json";
+import UserRegistry from "../../artifacts/contracts/UserRegistry.sol/UserRegistry.json";
 
 const userRegistryAddress = process.env.REACT_APP_DEPLOY_ADDRESS;
 
-const GetEntry = () => {
-  const [did, setDid] = useState("");
+const GetTotalEntries = () => {
   const [result, setResult] = useState("");
-
   // uses metamask injected browser window to make sure user has a connected account
   async function requestAccount() {
     await window.ethereum.request({ method: "eth_requestAccounts" });
   }
 
   // call to the insert method of the smart contract
-  async function getEntry() {
-    // making sure input is not empty
-    if (!did) {
-      console.log("GetEntry value for did is empty");
-      setResult(`GetEntry value for did is empty`);
-      return;
-    }
+  async function getTotalEntries() {
     if (typeof window.ethereum !== "undefined") {
       await requestAccount();
       const provider = new ethers.providers.Web3Provider(window.ethereum);
@@ -33,9 +25,9 @@ const GetEntry = () => {
         provider
       );
       try {
-        const data = await contract.getUser(did);
+        const data = await contract.getTotalEntries();
         console.log({ data });
-        setResult(`Retrieved Did: ${data[0]} with Key: ${data[1]}`);
+        setResult(`Total Entries: ${data}`);
       } catch (err) {
         console.log("Error: ", err);
         setResult("Error. Check console");
@@ -46,15 +38,9 @@ const GetEntry = () => {
   return (
     <div className="App">
       <header className="App-header">
-        <h2>Get Entry With User DID</h2>
-        <input
-          type="text"
-          required
-          placeholder="DID"
-          onChange={(e) => setDid(e.target.value)}
-        />
-        <button className="btn btn-outline-secondary" onClick={getEntry}>
-          Get Entry
+        <h2>Get Total Entries</h2>
+        <button className="btn btn-outline-secondary" onClick={getTotalEntries}>
+          Entry Count
         </button>
         {result}
       </header>
@@ -62,4 +48,4 @@ const GetEntry = () => {
   );
 };
 
-export default GetEntry;
+export default GetTotalEntries;
