+++ src/components/userRegistry/Insert.js
@@ -1,19 +1,22 @@
 import { useState } from "react";
 import { ethers } from "ethers";
-import "../css/bootstrap.css";
-import ConsumerRegistry from "../artifacts/contracts/ConsumerRegistry.sol/ConsumerRegistry.json";
+import "../../css/bootstrap.css";
+import QRCode from "qrcode";
+import QRCodeNew from "./qrcodeNew";
+import UserRegistry from "../../artifacts/contracts/UserRegistry.sol/UserRegistry.json";
 
-// May need to pdate on deployment. This is the address the contract is deployed to.\
-const consumerRegistryAddress = process.env.REACT_APP_CONSUMER_ADDRESS;
+// May need to update on deployment. This is the address the contract is deployed to.\
+const userRegistryAddress = process.env.REACT_APP_DEPLOY_ADDRESS;
 const verificationNum = process.env.REACT_APP_VERIFICATION;
 
-const InsertConsumer = () => {
+const Insert = () => {
   const [did, setDid] = useState("");
   const [contractKey, setContractKey] = useState("");
   const [result, setResult] = useState("");
   const [show, setShow] = useState(false);
+  const [src, setSrc] = useState("");
 
-  // uses metamask injected browser window to make sure consumer has a connected account
+  // uses metamask injected browser window to make sure user has a connected account
   async function requestAccount() {
     await window.ethereum.request({ method: "eth_requestAccounts" });
   }
@@ -49,10 +52,16 @@ const InsertConsumer = () => {
 
   // call to the insert method of the smart contract
   async function insert() {
-    // making sure input is not empty
-    if (!did || !contractKey) {
-      console.log("Insert values are empty");
-      setResult(`Insert Values are empty`);
+    // make sure input for did is not empty
+    if (!did) {
+      console.log("Insert value for did is empty");
+      setResult(`Insert value for did is empty`);
+      return;
+    }
+    // make sure input for contractKey is not empty
+    if (!contractKey) {
+      console.log("Insert value for contractKey is empty");
+      setResult(`Insert value for contractKey is empty`);
       return;
     }
 
@@ -78,8 +87,8 @@ const InsertConsumer = () => {
       // we need a signer because insert requires a transaction
       const signer = provider.getSigner();
       const contract = new ethers.Contract(
-        consumerRegistryAddress,
-        ConsumerRegistry.abi,
+        userRegistryAddress,
+        UserRegistry.abi,
         signer
       );
 
@@ -91,9 +100,12 @@ const InsertConsumer = () => {
       });
 
       try {
-        const transaction = await contract.insertConsumer(did, contractKey); //is there a way to get return value
+        const transaction = await contract.insertUser(did, contractKey); //is there a way to get return value
         // of non view function?
         await transaction.wait();
+        QRCode.toDataURL(did).then((data) => {
+          setSrc(data);
+        });
         console.log({ transaction });
       } catch (err) {
         console.log("Error: ", err);
@@ -105,7 +117,7 @@ const InsertConsumer = () => {
   return (
     <div className="App">
       <header className="App-header">
-        <h2>Insert Consumer to Consumer Registry</h2>
+        <h2>Insert User to User Regristry</h2>
         <input
           type="text"
           required
@@ -135,6 +147,7 @@ const InsertConsumer = () => {
               <div class="card-body">
                 <h5 class="card-title">Transaction Data</h5>
                 <p class="card-text">{result}</p>
+                <img src={src} />
               </div>
             </div>
           ) : null}
@@ -144,4 +157,4 @@ const InsertConsumer = () => {
   );
 };
 
-export default InsertConsumer;
+export default Insert;
