+++ src/components/userRegistry/qrcodeNew.js
@@ -1,0 +1,24 @@
+//TODO: Figure out how to rerender this component when the transaction
+// occurs. <QRCodeNew text={did}/>
+// Right now this code is inclused in Insert.js
+
+import QRCode from "qrcode";
+import { useEffect, useState } from "react";
+
+const QRCodeNew = ({ text }) => {
+  const [src, setSrc] = useState("");
+
+  useEffect(() => {
+    QRCode.toDataURL(text).then((data) => {
+      setSrc(data);
+    });
+  }, []);
+
+  return (
+    <div>
+      <img src={src} />
+    </div>
+  );
+};
+
+export default QRCodeNew;
